W_06/W06 Final Project Milestone/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
